Create Laxliife
This mod changes the SlowHand so that it can be used in place of the Grog and the Rubi. (Does not remove/change either)

Designed to Slag and Healthgate, this mod also makes the SlowHand fun to rocket jump (with a low level of course) and regenerate shotgun ammo. It also can mob decently well, having no trouble running through any map on UVHM+ extreme, but do note if used exclusively to mob, the SlowHand can run into ammo issues similar to that of the Interfacer. 

Add some life to your dedicated healing weapon!

This is my first Borderlands Mod, so feed back is appreciated and there's more stuff to come <3