Added 2 Guns
Guns Added:
BNK3R-9001  - A modified version of the Norfleet that fires Bunker Projectiles on Hero's Pass and on its own map. If you can figure out a way to extend this to all maps, please let me know. (This has a very high potential to crash your game or make you spawn in odd places. Just a warning ahead of time)
THE REVELATION - A variant of the storm that shoots Meteor Shower projectiles or Chain Lightning Projectiles depending on which one you switch it too. Also, Roulette Mode can be activated using the keys "9" and "Shift". If you do not hit the enemy at their feet just right, you'll die. If you hit them, it does a LOT of damage. You can press "9" to make the weapon not have any chance of killing you at all (applies to all weapons). There are two versions of this gun - THE REVELATION and THE REVELATIONNOP. THE REVELATION is an insta-kill weapon with these same attributes, while THE REVELATIONNOP is THE REVELATION but much more balanced (although still pretty strong)