Widowmaker
Replaces the Scorpio. Massively increased elemental chance and a DoT comparable to that of a grenade. Heals for 4 % of the damage inflicted. Becomes VERY accurate when aiming down sights.
Option to replace the bayonet mesh with the bandit version (enabled by default), alternatively moves the bayonet to accessory slot 2 (mind the sanity check!)