Add files via upload
This mod aims to make the game more like a bullet hell by slowing down the speed of enemy bullets. All guns used by enemies (including launcher enemies!) will have slower projectile speed than any of the player used counterparts. Weighting for rarities has been slightly adjusted, so legendary and purple guns will be a lot more common - but easier to dodge.. There are *6* options of speed to choose from, with a lowest of 50% and a highest of 300% decrease in bullet speed. The default is set to 200%. This mod requires the use of command extensions, so be sure to have that installed here: https://bl-sdk.github.io/mods/CommandExtensions/