Add files via upload
So this is should be a working version of Aaroon's Redux, Allegiance and Slayer COMs + their Legendary Versions as Skins to COMs.

That's 104 Legendary COMs added for Vanilla\UCP. 

Skins also should be easily editable to be whatever you want and, in my opinion redundant, Prefixes for COMs are freedup so you can name them whatever you want via Prefixes folder.

Was to lazy to check for conflicts with other stuff that reuses Skins, but probably should do that.

If anyone wants to grab it and repurpose or just use as is feel free. Especially since this might have typos i missed.