Update MakeVendorsGreatAgain And RefillVendors COMBOBREAKER
Missing brackets on two lines of code for shotgun and smg in both this file and the original. I dont know how to code, I just assume the brackets I replaced in the code are the correct ones causing the issue since they are the only two lines that dont look exactly like the others. Its a guess. going to try and fix it in my downloaded copy using blcmm and see if the errors go away.